verification que identifiant est un parametre defini

+++ malika/Front_end/output.c
@@ -1,3 +1,3 @@
-void renderUserProfile(int name) {
-	printf("<div> <h1> %d </h1> </div>", name);
+void renderUserProfile(float name) {
+	printf("<div> <h1> %f </h1> </div>", name);
 }
